Symmetries in Fourier space

From version 0.8, AtomicSymmetries.jl provided the possibility to apply symmetries to vector and matrices directly in Fourier space. This is implemented now for force-constant dynamical matrices and vectors (displacements, forces, ...).

A vector is transformed from real to q-space with the following convention:

\[\tilde v_a(\vec q) = \frac{1}{\sqrt{N_q}} \sum_{R} e^{-i 2\pi \vec q\cdot (\vec R + \vec\tau_a)} v_a(\vec R)\]

\[v_a(\vec R) = \frac{1}{\sqrt{N_q}} \sum_{\vec q} e^{i 2\pi \vec q\cdot(\vec R + \vec\tau_a)} \tilde v_a(\vec q)\]

where $\vec\tau_a$ is the position of the atom $a$ inside the primitive cell, so that $\vec R + \vec\tau_a$ is the equilibrium position of the atom in the supercell.

The phase factor employs the atomic positions $\vec R + \vec\tau_a$ and not only the origin $\vec R$ of the cell each atom belongs to. This is a gauge choice: the two conventions are related by the atom-dependent rephasing $e^{-2i\pi \vec q\cdot\vec\tau_a}$. The atomic-position gauge is adopted (since version 0.12) because it makes the Fourier transformed quantities smooth functions of $\vec q$, which is much better suited for interpolating quantities in q space. The price to pay is that quantities in this gauge are not periodic in the Brillouin zone:

\[\tilde v_a(\vec q + \vec G) = e^{-2i\pi \vec G\cdot \vec\tau_a}\, \tilde v_a(\vec q)\]

This must be carefully accounted for whenever a q point is folded back into the grid by a reciprocal lattice vector $\vec G$ (this occurs when applying symmetries, see below).

Note the sign of the Fourier and the normalization prefactor. This convention allows for correctly transforming the matrices, however, it introduces a size inconsistency on the vectors. If we have a periodic vector in the cell, its $q$ fourier transformed counterpart will be $\sqrt {N_q}$ times the value in the primitive cell. So be carefull when extracting $\Gamma$ point data from periodic vectors.

With this convention, we recover the standard rule for the matrices.

\[\tilde \Phi_{ab}(\vec q) = \sum_{\vec R} e^{2\pi i \vec q\cdot (\vec R + \vec\tau_a - \vec\tau_b)}\Phi_{a + \vec R;b}\]

\[\Phi_{ab} = \frac{1}{N_q} \sum_{\vec q} \tilde\Phi_{ab}(\vec q) e^{2i\pi \vec q\cdot[(\vec R(a) + \vec\tau_a) - (\vec R(b) + \vec\tau_b)]}\]

As for the vectors, the matrix in this gauge is not periodic in the Brillouin zone:

\[\tilde\Phi_{ab}(\vec q + \vec G) = e^{2i\pi \vec G\cdot(\vec\tau_a - \vec\tau_b)}\, \tilde\Phi_{ab}(\vec q)\]

Note that these transformation of matrices and vector are consistent so that matrices and vector written as outer product can be consistently transformed

\[\Phi(\vec R) = \sum_i\sum_{\vec R} \vec v_i(\vec R_1) \otimes \vec v_i(\vec R_1 + \vec R)\]

\[\tilde \Phi(\vec q) = \sum_i \vec {\tilde v}_i(\vec q) \otimes \vec {\tilde v_i}(-\vec q)\]

Notably, this convention introduces two main properties that must be handled with care. The $\Gamma$ value of the fourier transform is not the average over the supercell of the same quantity. If you want to obtain the average, you must divide by $\sqrt {N_q}$ (the number of q-points). If the coordinates passed as v_sc are absolute values of positions, then the $\Gamma$ value of the fourier transform will be shifted by a total translation which is the average of the equilibrium positions. This can be avoided by passing absolute_positions = true (which subtracts the equilibrium positions $\vec R + \vec\tau_a$ before transforming), or by removing this translational average a posteriori using the method shift_position_origin!.

Fourier transform

The API to perform the fourier transform occur mainly with vector_r2q!, vector_q2r! which, respectively, trasform a vector from real to q space and vice-versa. Transformation of matrices occur with matrix_r2q!, matrix_q2r!. All these operations are inplace. The matrices are assumed in crystal coordinates, but in this case it should not matter.

To shift the origin for the fourier transformed absolute positions, use the method shift_position_origin! as

The detailed API calls are

AtomicSymmetries.vector_r2q! — Function

vector_r2q!(
    v_q :: AbstractArray{Complex{T}, 3},
    v_sc :: AbstractMatrix{T},
    q :: Matrix{T},
    itau :: Vector{I},
    R_lat :: Matrix{T},
    tau :: Matrix{T}
) where {T <: AbstractFloat, I <: Integer}
vector_r2q!(v_q :: AbstractArray{Complex{T}, 2},
    v_sc :: AbstractVector{T},
    q :: Matrix{T},
    itau :: Vector{I},
    R_lat :: Matrix{T},
    tau :: Matrix{T}
) where {T <: AbstractFloat, I <: Integer}

Fourier transform a vector from real space and q space.

The phase-factor convention (gauge) includes the atomic positions $\vec \tau_a$ inside the primitive cell:

$\displaystyle v_a(\vec q) = \frac{1}{\sqrt{N_q}} \sum_{R} e^{-i 2\pi \vec q\cdot (\vec R + \vec\tau_a)} v_a(\vec R)$

where $\vec R + \vec \tau_a$ is the equilibrium position of the atom in the supercell. This gauge (in contrast with the one employing only $\vec R$) makes the Fourier transformed quantities smooth functions of $\vec q$, which is much better suited for interpolation in q space. Note that, as a consequence, quantities in this gauge are not periodic in the Brillouin zone: $v_a(\vec q + \vec G) = e^{-2i\pi \vec G\cdot \vec\tau_a} v_a(\vec q)$.

It works both on a single vector and on a series of vector. NOTE: In the latter case, the number of configurations must be in the first column. This is not standard, but implemented in this way for performance reasons as it is the most convenient memory rapresentation for vectorizing the average calculation.

Notably, this convention introduces two main properties:

The $\Gamma$ value of the fourier transform is not the average over the supercell of the same quantity. If you want to obtain the average, you must divide by √nq (the number of q-points).

If the coordinates passed as v_sc are absolute values of positions, then the $\Gamma$ value of the fourier transform will be shifted by a total translation proportional to the average of the equilibrium positions.

To avoid this behaviour (which is wrong), you can use the option absolute_positions, which automatically rescales the v_sc to be displacements with respect to the equilibrium positions $\vec R + \vec\tau_a$. Indeed, in this case, R_lat, tau and v_sc must be of the same units, and coordinate types (you cannot mix crystalline and cartesian).

Parameters

v_q : (nconfigs, 3nat, nq) The target vector in Fourier space. Optionally, nconfigs could be omitted if transforming only 1 vector

v_sc : (nconfigs, 3*natsc) The original vector in real space. Optionally, n_configs could be omitted if transforming only 1 vector

q : (3, nq) The list of q vectors

itau : (nat_sc) The correspondance for each atom in the supercell with the atom in the primitive cell.

R_lat : (3, nat_sc) The origin coordinates of the supercell in which the atom is

tau : (3, nat) The positions of the atoms inside the primitive cell. Must be expressed in the same units and coordinates as R_lat (and dual to q, so that $\vec q \cdot (\vec R + \vec\tau)$ is in units of 2π).

absolute_positions : Bool If true [default false], removes from vsc the equilibrium positions `Rlat + tau`.

AtomicSymmetries.vector_q2r! — Function

vector_q2r!(
    v_sc :: AbstractMatrix{T},
    v_q :: AbstractArray{Complex{T}, 3},
    q :: Matrix{T},
    itau :: Vector{I},
    R_lat :: Matrix{T},
    tau :: Matrix{T};
    absolute_positions :: Bool = false
    ) where {T <: AbstractFloat, I <: Integer}
function vector_q2r!(
    v_sc :: AbstractVector{T},
    v_q :: AbstractMatrix{Complex{T}},
    q :: Matrix{T},
    itau :: Vector{I},
    R_lat :: Matrix{T},
    tau :: Matrix{T};
    absolute_positions :: Bool = false
) where {T <: AbstractFloat, I <: Integer}

Fourier transform a vector from q space to real space.

The phase-factor convention (gauge) includes the atomic positions $\vec\tau_a$ inside the primitive cell (see vector_r2q!):

$\displaystyle v_a(\vec R) = \frac{1}{\sqrt{N_q}} \sum_{\vec q} e^{+i 2\pi \vec q\cdot (\vec R + \vec \tau_a)} v_a(\vec q)$

It can be applied both to a single vector and in an ensemble. NOTE: In the latter case, the configurations must be stored as the first index. This choice is made for performance reason in computing averages (exploiting vectorization).

Parameters

v_sc : (nconfigs, 3*natsc) The target vector in real space. Optionally, n_configs can be omitted

v_q : (nconfigs, nq, 3*nat) The original vector in Fourier space. Optionally, nconfigs can be omitted

q : (3, nq) The list of q vectors

itau : (nat_sc) The correspondance for each atom in the supercell with the atom in the primitive cell.

R_lat : (3, nat_sc) The origin coordinates of the supercell in which the atom is

tau : (3, nat) The positions of the atoms inside the primitive cell (same units and coordinates as R_lat).

absolute_positions : Bool If true, add the equilibrium positions R_lat + tau to the transformed v_sc

AtomicSymmetries.matrix_r2q! — Function

matrix_r2q!(
    matrix_q :: Array{Complex{T}, 3},
    matrix_r :: AbstractMatrix{T},
    q :: Matrix{T},
    itau :: Vector{I},
    R_lat :: Matrix{T},
    tau :: Matrix{T})

Fourier transform a matrix from real to q space

The phase-factor convention (gauge) includes the atomic positions $\vec\tau_a$ inside the primitive cell:

\[M_{ab}(\vec q) = \sum_{\vec R} e^{2\pi i \vec q\cdot (\vec R + \vec\tau_a - \vec\tau_b)}\Phi_{a + \vec R;b}\]

Where $\Phi_{ab}$ is the real space matrix, the $a+\vec R$ indicates the corresponding atom in the supercell displaced by $\vec R$. The phase is thus computed from the difference of the equilibrium positions of the two atoms, $(\vec R_a + \vec\tau_a) - (\vec R_b + \vec\tau_b)$. This gauge makes $M_{ab}(\vec q)$ a smooth function of $\vec q$, better suited for the interpolation. As a consequence, the matrix is not periodic in the Brillouin zone: $M_{ab}(\vec q + \vec G) = e^{2i\pi \vec G\cdot(\vec\tau_a - \vec\tau_b)} M_{ab}(\vec q)$.

Parameters

matrix_q : (3nat, 3nat, nq) The target matrix in Fourier space.

matrix_r : (3nat_sc, 3nat) The original matrix in real space (supercell)

q_tot : (3, nq) The list of q vectors

itau : (nat_sc) The correspondance for each atom in the supercell with the atom in the primitive cell.

R_lat : (3, nat_sc) The origin coordinates of the supercell in which the corresponding atom is

tau : (3, nat) The positions of the atoms inside the primitive cell (same units and coordinates as R_lat).

AtomicSymmetries.matrix_q2r! — Function

matrix_q2r!(
    matrix_r :: AbstractMatrix{T},
    matrix_q :: Array{Complex{T}, 3},
    q :: Matrix{T},
    itau :: Vector{Int},
    R_lat :: Matrix{T},
    tau :: Matrix{T})

Fourier transform a matrix from q space into r space

The phase-factor convention (gauge) includes the atomic positions $\vec\tau_a$ inside the primitive cell (see matrix_r2q!):

$\displaystyle \Phi_{ab} = \frac{1}{N_q} \sum_{\vec q} M_{ab}(\vec q) e^{2i\pi \vec q\cdot[(\vec R(a) + \vec\tau_a) - (\vec R(b) + \vec\tau_b)]}$

Where $\Phi_{ab}$ is the real space matrix, $M_{ab}(\vec q)$ is the q space matrix.

Parameters

matrixr : (3*natsc, 3*nat) The target matrix in real space (supercell). If the second dimension is 3nat_sc, we also apply the translations

matrix_q : (3nat, 3nat, nq) The original matrix in Fourier space.

q_tot : (3, nq) The list of q vectors

itau : (nat_sc) The correspondance for each atom in the supercell with the atom in the primitive cell.

Rlat : (3, natsc) The origin coordinates of the supercell in which the corresponding atom is

tau : (3, nat) The positions of the atoms inside the primitive cell (same units and coordinates as R_lat).

translations : Vector{Vector{Int}} The itau correspondance for each translational vector. Its size must be equal to the number of q point and contain all possible translations. This can be obtained from the get_translations subroutine.

AtomicSymmetries.shift_position_origin! — Function

shift_position_origin!(r_vector::AbstractVector{Complex{T}}, cell::AbstractMatrix{T}, R_lat::AbstractMatrix{T}; buffer) where T
shift_position_origin!(r_vectors::AbstractMatrix{Complex{T}}, cell::AbstractMatrix{T}, R_lat::AbstractMatrix{T}; buffer) where T

Shifts a set of position vectors (r_vector or r_vectors) such that the average position (center of mass) of the reference lattice (R_lat) is moved to the origin. The function operates in-place. This can be employed after performing a Fourier Transform of absolute positions, to correct for non centered R_lat.

This is typically used to remove the lattice translation from displacement vectors before calculating quantities that are invariant to rigid body translations.

The average position vector of the reference lattice $\mathbf{r}_{\text{avg}}$ is calculated in Cartesian coordinates, and then subtracted from the input vectors $\mathbf{r}$.

\[\mathbf{r}_{\text{avg}} = \frac{1}{N_{\text{atoms}}} \sum_{k=1}^{N_{\text{atoms}}} \mathbf{r}_{\text{lat}, k}\]

The position vectors are then shifted:

\[\mathbf{r}' = \mathbf{r} - \mathbf{r}_{\text{avg}}\]

Arguments

r_vector/r_vectors : The position vector(s) (of type Complex{T}) to be shifted in-place.

cell : The $N_{dims} \times N_{dims}$ matrix defining the unit cell (lattice vectors).

R_lat : The reference lattice positions (fractional/lattice coordinates, $N_{dims} \times N_{atoms}$) whose average position determines the shift.

buffer : An optional buffer for temporary memory allocation (Bumer.jl)

Details on r_vectors Matrix Method

The matrix method assumes that the input r_vectors has the dimensions $(N_{\text{configs}}, N_{\text{dims}} \times N_{\text{atoms,v}})$, where $N_{\text{configs}}$ is the number of configurations, and $N_{\text{atoms,v}}$ is the number of atoms in the vectors being shifted.

The shift is applied simultaneously to all configurations using broadcasting (@views ... .-=).

Example

An example of usage after the fourier transform

# Define positions_r as a n_configs, 3n_atoms_sc vector
# Perform the fourier transform in q space
# Here, we assume that R_lat and q_points are expressed in crystal coordinates.
# Otherwise, just pass the identity to the cell below.
vector_r2q!(positions_q, positions_r, q_points, itau, R_lat, tau)

# Remove the translations
shift_position_origin!(positions_q, cell, R_lat)

Symmetries in Q space

The application of symmetries in Fourier space must also account how points in q space are mapped by the symmetry operations.

For this, the important information about how q points are related by symmetries needs to be computed and stored. This identification is performed by the helper function get_irt_q!, which identifies, for a given symmetry operation, the i->j mapping between q points. Q points mapped into themselves by the same set of symmetry operations form the socalled small-group of $q$, while the set of $q$ points mapped by all the symmetries of a crystal is called the star of the $q$ point. Due to time-inversion symmetry, the dynamical matrix must also satisfy the condition

$D(q) = D^\dagger(-q + G)$

(where, in the atomic-position phase gauge, the folding by $G$ introduces the block phase factor $e^{2i\pi \vec G\cdot(\vec\tau_a - \vec\tau_b)}$, see impose_hermitianity_q!), therefore it is necessary also to keep track, for each q point, which one is the corresponding $-q + G$ in the mesh. This mapping is computed by the helper function get_minus_q!. All these information needs to be stored when applying symmetries. Therefore we defined a new Symmetries struct that ihnerits from the GenericSymmetries called SymmetriesQSpace. Note that, to initialize the symmetries in q-space, we must use the symmetries object (Symmetries) evaluated in the primitive cell. The correct initialization of symmetries could be checked with the subroutine check_symmetries, which will spot if a different cell has been employed when initializing the symmetries.

Since the q points must be passed in crystal coordinates, it may be useful to get the reciprocal lattice, which can be done with $get_reciprocal_lattice!$ (see section on crystal coordinates for the API)

Application of symmetries

Applying a symmetry means transforming a vector or a matrix (already in q-space) into a new vector (matrix). If the vector (matrix) is invariant under that transformation, then that transformation belong to the symmetry group.

Notably, the symmetries in the supercell are always the symmetries in the primitive cell times all possible translations operated by the lattice vectors compatible with the chosen supercell. On the contrary, the symmetries in q space are always only equal to the symmetries in the primitive cell. The reason is that translations are automatically incorporated in the q space representation by the block theorem:

\[D(q, q') = D(q)\delta(q - q')\]

This means that applying each symmetry operation in $q$ space is equivalent to averaging the result of the same symmetry operation in the supercell averaging among all possible translations.

The application of a symmetry in q space can be performed by considering how the force-constant matrix transforms in real space under a symmetry operation $\{S | \vec v\}$ (rotation $S$ plus fractional translation $\vec v$, in crystal coordinates).

In the atomic-position phase gauge the derivation is particularly simple. The symmetry maps the equilibrium position of the atom $a$ in the cell $\vec R$ into the equilibrium position of the atom $s(a)$ in another cell:

\[S(\vec R + \vec\tau_a) + \vec v = \vec R\,' + \vec \tau_{s(a)}\]

Since the phase factors of the Fourier transform are computed exactly from these equilibrium positions, the phases follow the atoms through the symmetry operation, and the transformation of the dynamical matrix takes the form

\[S[\tilde\Phi_{ab}(\bm q)] = S^\dagger\, \tilde\Phi_{s(a)s(b)}(S_\text{recip}\bm q)\, S, \qquad \bm S_\text{recip} = \left(\bm S^{-1}\right)^\dagger\]

with no phase factor associated with the fractional translations: the phases $e^{-2i\pi (S_\text{recip}\bm q)\cdot \vec v}$ picked up by the two displacement vectors cancel between the two atomic indices (this is one of the advantages of this gauge with respect to the lattice one, where a phase factor $e^{2\pi i \bm q\cdot(\bm t_{s,a} - \bm t_{s,b})}$ involving the unit-cell translations $\bm t_{s,a} = S\vec\tau_a + \vec v - \vec\tau_{s(a)}$ appears).

However, in this gauge the dynamical matrix is not periodic in the reciprocal lattice. The vector $S_\text{recip}\bm q$ may fall outside the q grid, and it is folded back into the grid point $\bm q_{\text{grid}}$ by a reciprocal lattice vector $\bm G$:

\[S_\text{recip}\bm q = \bm q_{\text{grid}} + \bm G\]

The folding introduces the phase factor

\[\tilde\Phi_{s(a)s(b)}(\bm q_\text{grid} + \bm G) = e^{2\pi i \bm G\cdot(\vec\tau_{s(a)} - \vec\tau_{s(b)})}\, \tilde\Phi_{s(a)s(b)}(\bm q_\text{grid})\]

which is computed and applied automatically by apply_symmetry_matrixq!. For this reason, the atomic positions inside the primitive cell (in crystal coordinates) must be provided when initializing SymmetriesQSpace, and they must be the same positions employed in the Fourier transform (choosing a different periodic image of an atom changes the gauge).

For vectors, the fractional translation phase does not cancel, and the transformation reads

\[S[\tilde v_{a}(\bm q)] = e^{-2\pi i (S_\text{recip}\bm q)\cdot \vec v}\, e^{2\pi i \bm G\cdot \vec\tau_{s(a)}}\, S\, \tilde v_{a}(\bm q)\]

These phases are applied by apply_symmetry_vectorq! when the positions, the q points and the fractional translation are provided (they are automatically provided when using the general interface rotate_vector!).

The application of symmetries is handled by the general function rotate_vector! and rotate_dynamical_matrix! or rotate_matrix! that works exactly like for real space symmetries, with the same general interface. However, we also provide specific q-space only functions. Note that, while the rotate_* functions works in cartesian space, the following one expects symmetries in real space.

This transformation for each q point is operated by the subroutine apply_symmetry_matrixq!. Both these function modify in-place the first argument, storing the result of the transformation there. Note that, since symmetries are stored in crystalline components, both the vector and the matrix must be in crystalline components. This makes it also important that the $\bm q$ points are provided in crystalline coordinates, to correctly compute the phase factor and the transformed $S\bm q$.

AtomicSymmetries.SymmetriesQSpace — Type

SymmetriesQSpace(symmetries :: Symmetries{T}, q_points :: AbstractMatrix{T}, positions :: AbstractMatrix{T}) :: SymmetriesQSpace{T} where T

struct SymmetriesQSpace{T} <: GenericSymmetries where T
    symmetries :: Symmetries{T}
    q_points :: Matrix{T}
    positions :: Matrix{T}
    irt_q :: Vector{Vector{Int}}
    minus_q_index :: Vector{Int}
end

This structure contains the information to perform the symmetrization of a dynamical matrix directly in q space.

Note that the q_points and the positions need to be in crystal coordinates,

and the symmetries must be of the primitive cell.

The atomic positions are required because the Fourier transform is defined with the phase factor gauge $e^{2i\pi \vec q\cdot(\vec R + \vec\tau_a)}$ (see vector_r2q! and matrix_r2q!). In this gauge the q-space quantities are not periodic in the Brillouin zone: whenever a symmetry maps a q point outside the grid ($S^{-T}\vec q = \vec q' + \vec G$), the folding by the reciprocal lattice vector $\vec G$ introduces phase factors depending on $\vec G\cdot\vec\tau_a$.

Important: the positions must be the same used to compute the Fourier transform (expressed in crystal coordinates); using a different periodic image of an atom changes the gauge.

Parameters

symmetries : The symmetries of the primitive cell (Symmetries{T})

q_points : The q points where the symmetries must be applied (in crystal coordinates)

positions : The atomic positions in the primitive cell (in crystal coordinates); size (ndims, natoms)

irt_q : A vector (one for each symmetry) of the correspondances of q points. For each symmetry can be obtained from get_irt_q! Q points linked in this way are related by the symmetry operation in reciprocal space, and belong to the same star of q.

minus_q_index : A vector containing for each q the corresponding $\vec {q'} = -\vec q + \vec G$, where $\vec G$ is a generic reciprocal lattice vector.

AtomicSymmetries.apply_symmetry_vectorq! — Function

apply_symmetry_vectorq!(target_vector :: AbstractMatrix{Complex{T}}, original_vector :: AbstractMatrix{Complex{T}}, symmetry_operation :: AbstractMatrix{U}, irt :: Vector{Int}, irt_q:: AbstractVector{Int}; positions = nothing, q_points = nothing, translation = nothing, buffer = default_buffer())

Apply the symmetry on the original vector in q space.

The vector in q space is assumed in the phase factor gauge $e^{-2i\pi \vec q\cdot(\vec R + \vec\tau_a)}$ (see vector_r2q!). In this gauge the application of a symmetry $\{S | \vec v\}$ reads

\[v'_{s(a)}(S^{-T}\vec q) = e^{-2i\pi (S^{-T}\vec q)\cdot \vec v}\, S\, v_a(\vec q)\]

with no atom-dependent phase. However, when $S^{-T}\vec q$ falls outside the q grid, it is folded back by a reciprocal lattice vector $\vec G$ ($S^{-T}\vec q = \vec q_{\text{grid}} + \vec G$), and, since in this gauge vectors are not periodic in the Brillouin zone, the folding introduces the extra phase $e^{2i\pi \vec G\cdot \vec\tau_{s(a)}}$.

If positions and q_points are provided (in crystal coordinates), all these phase factors are computed and applied. If they are omitted (default), no phase is applied: this is only correct if the result is used at $\Gamma$ (e.g. inside symmetrize_vector_q!), or when all the phases are trivial (symmorphic group, no folding).

Parameters

target_vector : The result (modified inplace) (3n x nq)

original_vector : The original vector (3n x nq)

symmetry_operation : The 3x3 symmetry (crystal coordinates, direct space)

irt : The atom-atom association by symmetry

irt_q : The q-q association by symmetry

positions : The atomic positions in the primitive cell (crystal coordinates, ndims x natoms) [Optional, keyword]

q_points : The q points (crystal coordinates, n_dims x nq) [Optional, keyword]

translation : The fractional translation $\vec v$ of the symmetry (crystal coordinates) [Optional, keyword]

buffer : The Bumper.jl buffer for caching memory allocations [Optional, keyword]

AtomicSymmetries.apply_symmetry_matrixq! — Function

apply_symmetry_matrixq!(target_matrix :: AbstractArray{Complex{T}, 3},
    original_matrix :: AbstractArray{Complex{T}, 3},
    sym :: AbstractMatrix{U},
    irt :: AbstractVector{Int},
    irt_q :: AbstractVector{Int},
    positions :: AbstractMatrix{T},
    q_points :: AbstractMatrix{T}
    ; buffer = default_buffer()) where {T, U}

Apply the symmetry on the matrix in q space.

The matrix in q space is assumed in the phase factor gauge $e^{2i\pi \vec q\cdot(\vec R + \vec\tau_a)}$ (see matrix_r2q!), where the phase factors are computed from the atomic positions, not from the origin of the supercell in which each atom lies.

In this gauge, the application of a symmetry $\{S | \vec v\}$ reads

\[M'_{s(a)s(b)}(S^{-T}\vec q) = S\, M_{ab}(\vec q)\, S^T\]

with no phase factors coming from the fractional translations (they cancel between the two atoms). However, in this gauge the matrix is not periodic in the Brillouin zone: whenever $S^{-T}\vec q$ falls outside the q grid and is folded back by a reciprocal lattice vector $\vec G$ ($S^{-T}\vec q = \vec q_{\text{grid}} + \vec G$), the folding introduces the phase factor $e^{2i\pi \vec G\cdot(\vec\tau_{s(a)} - \vec\tau_{s(b)})}$, which is accounted for by this subroutine. Everything (symmetries, positions and q points) must be provided in crystal coordinates.

Parameters

target_matrix : The result (modified inplace) (3n x 3n x nq)

original_matrix : The original matrix (3n x 3n x nq)

sym : The 3x3 symmetry (crystal coordinates, direct space)

irt : The atom-atom association by symmetry

irt_q : The q-q association by symmetry

positions : The atomic positions inside the primitive cell (crystal coordinates, ndims x natoms). They must be the same positions used to compute the Fourier transform.

q_points : The q points (crystal coordinates, n_dims x nq)

buffer : The Bumper.jl buffer for caching memory allocations [Optional]

AtomicSymmetries.get_irt_q! — Function

get_irt_q!(irt_q :: AbstractVector{Int}, q_points :: AbstractVector{T}, sym_mat :: AbstractMatrix)

Get the correspondance $q' = S_\text{recip} q$ on the provided q grid. Always assume everything is in crystal coordinates.

Note that in reciprocal space (crystal coordinates) the symmetry operation is the inverse transpose.

\[S_\text{recip} = (S_\text{direct})^{-T}\]

The provided sym_mat is assumed to be in direct space.

This is needed for the correct application of the symmetries

AtomicSymmetries.get_minus_q! — Function

get_minus_q!(minus_q_index :: AbstractVector{Int}, q_points :: AbstractMatrix{T}; buffer = default_buffer()) where T
get_minus_q!(minus_q_index :: AbstractVector{Int}, q_points :: AbstractMatrix{T}, reciprocal_lattice :: AbstractMatrix{T}; buffer = default_buffer()) where T

Identify for each q point what is the corresponding -q:

$\vec q \longrightarrow -\vec q + \vec G$

where $\vec G$ is a reciprocal vector. Since this is done in crystal coordinates$\vec G$ are all possible integers.

The reciprocal vectors is only needed if the q points are in cartesian coordinates.

Parameters

minus_q_index : The result vector (modified inplace)

q_points : The q points (in crystal coordinates if no reciprocal_lattice is provided, cartesian otherwise)

cell : The primitive cell (column-based). Only if q_points are in cartesian coordinates. [Optional]

reciprocal_lattice : The reciprocal lattice vectors (column-based). Only if q_points are in cartesian coordinates. [Optional]

buffer : The Bumper.jl buffer for caching memory allocations [Optional, keyword only]

AtomicSymmetries.check_symmetries — Function

check_symmetries(q_symmetries :: SymmetriesQSpace{T}, n_atoms :: Int) :: Bool

Check if the q_symmetries has been correctly initialized in the primitive cell.

Essentially, this subroutine checks the atomic correspondance by symmetry and spots if there are atoms outside the primitive cell (whose index is above n_atoms).

Parameters

q_symmetries : The symmetries in q space

n_atoms : The number of atoms in the primitive cell

Returns

true if no contraddiction have been detected, false otherwise.

Enforcing symmetries

One of the most useful operation to do is enforce a specific matrix or vector in q-space to satisfy a given symmetry group.

This can be implemented by applying the complete irreducible representation of the symmetry group. Symmetrization of an ent \Phi is obtained as

$\Phi = \frac{1}{N}\sum_{i=1}^N S_i(\Phi)$

where $S_i$ is the symmetry operation. The two functions performing the symmetrization are symmetrize_matrix_q! and symmetrize_vector_q!. Also in this case, the dynamical matrix must be provided in crystalline coordinates.

To symmetrize vector and matrices already provided in cartesian coordinates, we must use the appropriate subroutines symmetrize_vector_cartesian_q! and symmetrize_matrix_cartesian_q!. These two subroutines correctly convert the vector/matrix in crystal coordinates before applying the symmetries, and then convert the symmetrized result back in cartesian space. They are the most used subroutines to perform symmetrization in q-space, the equivalent of symmetrize_vector! and symmetrize_fc! for real space.

The Hermitianity is not automatically imposed by the symmetrization procedure. This allows to symmetrize matrices that are not necessarily hermitian, for example, the cross correlation matrices between different quantities. Hermitianity and time-reversal symmetry can be imposed with the subroutine impose_hermitianity_q!, which enforces the condition. The time-reversal symmetry corresponds to the condition that the original matrix in real space is real-valued.

Here the complete API

AtomicSymmetries.symmetrize_vector_q! — Function

symmetrize_vector_q!(target_gamma :: AbstractVector{T}, original_q :: AbstractArray{Complex{T}, 2}, symmetries :: Symmetries, irt_q :: Vector{Vector{Int}}; buffer = default_buffer() where T

Impose the symmetrization of a vector in q space. Since the symmetrization also imposes translational symmetries, the result is always a vector only at gamma.

The symmetrized vector is supposed to be a displacement (so no translations are applied)

NOTE: The provided vector must be in crystal coordinates To symmetrize a vector incartesian coordinates, see the routine symmetrize_vector_cartesian_q!.

Parameters

target_gamma : The n_at * n_dims output symmetrized vector at $\Gamma$

original_q : The original vector in q-space of size nat*n_dims, nq

symmetries : The symmetry group

irt_q : A vector (one for each symmetry) of the correspondances of q points. For each symmetry can be obtained from get_irt_q!

gamma_index : Specify which q vector is $\Gamma$. If not specified, it is assumed to be the first one

AtomicSymmetries.symmetrize_matrix_q! — Function

symmetrize_matrix_q!(target_q :: AbstractArray{Complex{T}, 3}, original_q :: AbstractArray{Complex{T}, 3}, symmetries :: Symmetries, irt_q :: Vector{Vector{Int}}; buffer = default_buffer() where T
symmetrize_matrix_q!(matrix_q :: AbstractArray{Complex{T}, 3}, q_symmetries :: SymmetriesQSpace; buffer = default_buffer())  where T
symmetrize_matrix_q!(target_q :: AbstractArray{T, 3}, original_q :: AbstractArray{Complex{T}, 3}, q_symmetries :: SymmetriesQSpace; buffer = default_buffer())  where T

Impose the symmetrization of a dynamical matrix in q space. The matrix must be in crystal coordinates.

Parameters

target_q : The symmetrized matrix of size n_modes, n_modes, nq (modified in-place).

original_q : The original matrix in q-space of size n_modes, n_modes, nq. It could be the same as target_q

symmetries : The symmetry group

irt_q : A vector (one for each symmetry) of the correspondances of q points. For each symmetry can be obtained from get_irt_q!

positions : The atomic positions in the primitive cell (crystal coordinates, ndims x natoms). They must be the same positions used to compute the Fourier transform (see apply_symmetry_matrixq!).

minus_q_index : A vector containing for each q the corresponding $\vec {q'} = -\vec q + \vec G$, where $\vec G$ is a generic reciprocal lattice vector.

q_points : The vector containing the actual q points

AtomicSymmetries.symmetrize_vector_cartesian_q! — Function

symmetrize_vector_cartesian_q!(vector_q_cart:: AbstractArray{Complex{T}, 2}, cell :: Matrix{T}, symmetries :: SymmetriesQSpace; buffer = default_buffer()) where {T}

Perform the symmetrization of a vector (overwriting it) in cartesian coordinates. This is the go-to subroutine for performing symmetrization of vectors in q space.

Parameters

vector_q_cart : in-place symmetrize vector (q-space, cartesian coordinates)

cell : 3x3 matrix of the primitive cell (column-based)

symmetries : Symmetries in Q space

buffer : Optional, Bumper stack buffer (caching)

AtomicSymmetries.symmetrize_matrix_cartesian_q! — Function

symmetrize_matrix_cartesian_q!(matrix_q :: AbstractArray{Complex{T}, 3}, cell :: Matrix{T}, q_symmetries :: SymmetriesQSpace; buffer=default_buffer()) where T

Enforce the symmetries on the provided matrix (q-space), modifying it in-place. The provided matrix must be in Cartesian Coordinates.

Parameters

matrix_q : The matrix to be symmetrized. Size (nmodes, nmodes, nq)

cell : The 3x3 primitive cell (column ordering)

q_symmetries : The symmetry group (q-space)

buffer : Optional, stack for Bumper to cache allocations.

AtomicSymmetries.impose_hermitianity_q! — Function

impose_hermitianity_q!(matrix_q :: AbstractArray{Complex{T}, 3}, minus_q_index, positions :: AbstractMatrix{T}, q_points :: AbstractMatrix{T}; buffer=default_buffer()) where T
impose_hermitianity_q!(matrix_q :: AbstractArray{Complex{T}, 3}, q_symmetries :: SymmetriesQSpace{T}; buffer=default_buffer()) where T
impose_hermitianity_q!(matrix_q :: AbstractArray{Complex{T}, 3}, minus_q_index; buffer=default_buffer()) where T

Impose the hermitianity and time-reversal symmetry on the dynamical matrix in q space.

The matrix in q space is assumed in the phase factor gauge $e^{2i\pi \vec q\cdot(\vec R + \vec\tau_a)}$ (see matrix_r2q!). In this gauge the time-reversal constraint $M_{ab}(-\vec q) = M_{ab}(\vec q)^*$ still holds exactly; however, since $-\vec q$ is stored in the grid at $\vec q_j = -\vec q + \vec G_m$, and the matrix is not periodic in the Brillouin zone, the folding introduces the block phase factor $e^{2i\pi \vec G_m\cdot(\vec\tau_a - \vec\tau_b)}$. For this reason the atomic positions and the q_points (both in crystal coordinates) are needed.

The method accepting only minus_q_index does not apply any folding phase: it is correct only when all the phases are trivial (e.g. all the atoms sit at the origin of the cell, or the q grid is closed under inversion without folding, like a grid centered around $\Gamma$).

Parameters

matrix_q : The dynamical matrix in q space (modified inplace)

minus_q_index : A vector containing for each q the corresponding $\vec {q'} = -\vec q + \vec G$, where $\vec G$ is a generic reciprocal lattice vector.

positions : The atomic positions in the primitive cell (crystal coordinates, ndims x natoms). They must be the same positions used to compute the Fourier transform.

q_points : The q points (crystal coordinates, n_dims x nq)

q_symmetries : The symmetries in q space (contains minus_q_index, positions and q_points)

Manipulating q points

The fourier transform depends on the knowledge of few vectors: q_points, itau, and R_lat (evenutally translations, for inverse fourier transform into a matrix).

All these properties can be evaluated from the core source. For example, to obtain the supercell to which the q points are commensurate, we can use the get_supercell method as

AtomicSymmetries.get_supercell! — Function

get_supercell(q_points::AbstractMatrix{T}, cell :: AbstractMatrix{T}) :: Vector{Int}
get_supercell(q_points::AbstractMatrix{T}) :: Vector{Int}
get_supercell!(supercell::AbstractVector{I}, q_points::AbstractMatrix{T}, cell :: AbstractMatrix{T}) where {T, I<:Integer}
get_supercell!(supercell::AbstractVector{I}, q_points::AbstractMatrix{T}) where {T, I<:Integer}

Calculates the minimum supercell dimensions required to fold a set of commensurate wave vectors (q_points) back to the Gamma point (Γ) of the Brillouin zone.

The resulting supercell dimension $S_i$ for each spatial direction is determined by the reciprocal of the smallest non-zero q-point component in that direction. For commensurate grids, this is mathematically equivalent to:

\[S_i = \text{round} \left( \frac{1}{\min(|q_{i}|)} \right)\]

This function is primarily used when performing calculations in a real-space supercell that is commensurate with the input k-point (or q-point) grid.

Arguments

q_points: A 2D matrix where each column represents a q-point vector, and each row corresponds to a dimension (x, y, z).

supercell: An pre-allocated integer vector to store the result (used by the get_supercell! in-place version).

cell : The primitive cell. If not provided, the code assumes that q_points are in fractional coordinates.

Important Note on Coordinates

The q_points must be provided in crystal coordinates (fractional coordinates) if cell is not provided.

Example

# 3 dimensions, 2 q-points
q_points = [0.5 0.0;
            0.0 0.5;
            0.0 0.25]

# The supercell dimensions required are based on (1/0.5, 1/0.5, 1/0.25).
supercell = get_supercell(q_points)
# Result: [2, 2, 4]

Analogously, we can get the translations R_lat as

AtomicSymmetries.get_R_lat! — Function

get_R_lat!(R_lat :: Matrix{T}, primitive_coords :: Matrix{T}, supercell_coords :: Matrix{T})

Get the R_lat parameter to perform the fourier transform.

Parameters

R_lat the result lattice vectors, modified inplace

primitive_coords : The coordinates in the primitive cell

supercell_coords : The coordinates in the supercell

itau : The correspondence for each atom in the supercell with the respective atom in the primitive cell