Symmetries in Fourier space

From version 0.8, AtomicSymmetries.jl provided the possibility to apply symmetries to vector and matrices directly in Fourier space. This is implemented now for force-constant dynamical matrices and vectors (displacements, forces, ...).

A vector is transformed from real to q-space with the following convention:

$\displaystyle \tilde v_k(\vec q) = \frac{1}{\sqrt{N_q}} \sum_{R} e^{-i 2\pi \vec R\cdot \vec q} v_k(\vec R)$

$\displaystyle v_k(\vec R) = \frac{1}{\sqrt{N_q}} \sum_{R} e^{i 2\pi \vec R\cdot \vec q} \tilde v_k(\vec q)$

Note the sign of the Fourier and the normalization prefactor. With this convention, we recover the standard rule for the matrices.

$\displaystyle \tilde \Phi_{ab}(\vec q) = \sum_{\vec R} e^{2\pi i \vec q\cdot \vec R}\Phi_{a;b + \vec R}$

$\displaystyle \Phi_{ab} = \frac{1}{N_q} \sum_{\vec q} \tilde\Phi_{ab}(\vec q) e^{2i\pi \vec q\cdot[\vec R(a) - \vec R(b)]}$

Note that these transformation of matrices and vector are consistent so that matrices and vector written as outer product can be consistently transformed

$\displaystyle \Phi(\vec R) = \sum_i\sum_{\vec R} \vec v_i(\vec R_1) \otimes \vec v_i(\vec R_1 + \vec R)$

$\displaystyle \tilde \Phi(\vec q) = \sum_i \vec {\tilde v}_i(\vec q) \otimes \vec {\tilde v_i}(-\vec q)$

Fourier transform

The API to perform the fourier transform occur mainly with vector_r2q!, vector_q2r! which, respectively, trasform a vector from real to q space and vice-versa. Transformation of matrices occur with matrix_r2q!, matrix_q2r!. All these operations are inplace. The matrices are assumed in crystal coordinates, but in this case it should not matter.

The detailed API calls are

AtomicSymmetries.vector_r2q! — Function

vector_r2q!(
    v_q :: AbstractArray{Complex{T}, 3},
    v_sc :: AbstractMatrix{T},
    q_tot :: Matrix{T})
vector_r2q!(v_q :: AbstractArray{Complex{T}, 2},
    v_sc :: AbstractVector{T},
    q :: Matrix{T},
    itau :: Vector{I},
    R_lat :: Matrix{T}
) where {T <: AbstractFloat, I <: Integer}

Fourier transform a vector from real space and q space.

$\displaystyle v_k(\vec q) = \frac{1}{\sqrt{N_q}} \sum_{R} e^{-i 2\pi \vec R\cdot \vec q} v_k(\vec R)$

It works both on a single vector and on a series of vector. NOTE: In the latter case, the number of configurations must be in the first column. This is not standard, but implemented in this way for performance reasons as it is the most convenient memory rapresentation for vectorizing the average calculation.

Parameters

v_q : (nconfigs, 3nat, nq) The target vector in Fourier space. Optionally, nconfigs could be omitted if transforming only 1 vector

v_sc : (nconfigs, 3*natsc) The original vector in real space. Optionally, n_configs could be omitted if transforming only 1 vector

q_tot : (3, nq) The list of q vectors

itau : (nat_sc) The correspondance for each atom in the supercell with the atom in the primitive cell.

R_lat : (3, nat_sc) The origin coordinates of the supercell in which the atom is

AtomicSymmetries.vector_q2r! — Function

vector_q2r!(
    v_sc :: AbstractMatrix{T},
    v_q :: AbstractArray{Complex{T}, 3},
    q_tot :: Matrix{T},
    itau :: Vector{I},
    R_lat :: Matrix{T}) where {T <: AbstractFloat, I <: Integer}
function vector_q2r!(
    v_sc :: AbstractVector{T},
    v_q :: AbstractMatrix{Complex{T}},
    q :: Matrix{T},
    itau :: Vector{I},
    R_lat :: Matrix{T}
) where {T <: AbstractFloat, I <: Integer}

Fourier transform a vector from q space to real space.

$\displaystyle v_k(\vec R) = \frac{1}{\sqrt{N_q}} \sum_{R} e^{+i 2\pi \vec R\cdot \vec q} v_k(\vec q)$

It can be applied both to a single vector and in an ensemble. NOTE: In the latter case, the configurations must be stored as the first index. This choice is made for performance reason in computing averages (exploiting vectorization).

Parameters

v_sc : (nconfigs, 3*natsc) The target vector in real space. Optionally, n_configs can be omitted

v_q : (nconfigs, nq, 3*nat) The original vector in Fourier space. Optionally, nconfigs can be omitted

q_tot : (3, nq) The list of q vectors

itau : (nat_sc) The correspondance for each atom in the supercell with the atom in the primitive cell.

R_lat : (3, nat_sc) The origin coordinates of the supercell in which the atom is

AtomicSymmetries.matrix_r2q! — Function

matrix_r2q!(
    matrix_q :: Array{Complex{T}, 3},
    matrix_r :: AbstractMatrix{T},
    q :: Matrix{T},
    itau :: Vector{I},
    R_lat :: Matrix{T})

Fourier transform a matrix from real to q space

\[M_{ab}(\vec q) = \sum_{\vec R} e^{-2\pi i \vec q\cdot \vec R}\Phi_{a;b + \vec R}\]

Where $\Phi_{ab}$ is the real space matrix, the $b+\vec R$ indicates the corresponding atom in the supercell displaced by $\vec R$.

Parameters

matrix_q : (3nat, 3nat, nq) The target matrix in Fourier space.

matrix_r : (3nat_sc, 3nat) The original matrix in real space (supercell)

q_tot : (3, nq) The list of q vectors

itau : (nat_sc) The correspondance for each atom in the supercell with the atom in the primitive cell.

R_lat : (3, nat_sc) The origin coordinates of the supercell in which the corresponding atom is

AtomicSymmetries.matrix_q2r! — Function

matrix_q2r!(
    matrix_r :: AbstractMatrix{T},
    matrix_q :: Array{Complex{T}, 3},
    q :: Matrix{T},
    itau :: Vector{Int},
    R_lat :: Matrix{T})

Fourier transform a matrix from q space into r space

$\displaystyle \Phi_{ab} = \frac{1}{N_q} \sum_{\vec q} M_{ab}(\vec q) e^{2i\pi \vec q\cdot[\vec R(a) - \vec R(b)]}$

Where $\Phi_{ab}$ is the real space matrix, $M_{ab}(\vec q)$ is the q space matrix.

Parameters

matrixr : (3*natsc, 3*nat) The target matrix in real space (supercell). If the second dimension is 3nat_sc, we also apply the translations

matrix_q : (3nat, 3nat, nq) The original matrix in Fourier space.

q_tot : (3, nq) The list of q vectors

itau : (nat_sc) The correspondance for each atom in the supercell with the atom in the primitive cell.

Rlat : (3, natsc) The origin coordinates of the supercell in which the corresponding atom is

translations : Vector{Vector{Int}} The itau correspondance for each translational vector. Its size must be equal to the number of q point and contain all possible translations. This can be obtained from the get_translations subroutine.

Applications of symmetries in Fourier space

The application of symmetries in Fourier space must also account how points in q space are mapped by the symmetry operations.

For this, the important information about how q points are related by symmetries needs to be computed and stored. This identification is performed by the helper function get_irt_q!, which identifies, for a given symmetry operation, the i->j mapping between q points. Q points mapped into themselves by the same set of symmetry operations form the socalled small-group of $q$, while the set of $q$ points mapped by all the symmetries of a crystal is called the star of the $q$ point. Due to time-inversion symmetry, the dynamical matrix must also satisfy the condition

$D(q) = D^\dagger(-q + G)$

therefore it is necessary also to keep track, for each q point, which one is the corresponding $-q + G$ in the mesh. This mapping is computed by the helper function get_minus_q!. All these information needs to be stored when applying symmetries. Therefore we defined a new Symmetries struct that ihnerits from the GenericSymmetries called SymmetriesQSpace

AtomicSymmetries.SymmetriesQSpace — Type

SymmetriesQSpace(symmetries :: Symmetries{T}, q_points :: AbstractMatrix{T}) :: SymmetriesQSpace{T} where T

struct SymmetriesQSpace{T} <: GenericSymmetries where T
    symmetries :: Symmetries{T}
    irt_q :: Vector{Vector{Int}}
    minus_q_index :: Vector{Int}
end

This structure contains the information to perform the symmetrization of a dynamical matrix directly in q space. Note that the q_points needs to be in crystal coordinates.

AtomicSymmetries.get_irt_q! — Function

get_irt_q!(irt_q :: AbstractVector{Int}, q_points :: AbstractVector{T}, sym_mat :: AbstractMatrix)

Get the correspondance q' = Sq on the provided q grid. Always assume everything is in crystal coordinates

AtomicSymmetries.get_minus_q! — Function

get_minus_q!(minus_q_index :: AbstractVector{Int}, q_points :: AbstractMatrix{T}; buffer = default_buffer()) where T

Identify for each q point what is the corresponding -q:

$\vec q \longrightarrow -\vec q + \vec G$

where $\vec G$ is a reciprocal vector. Since this is done in crystal coordinates$\vec G$ are all possible integers.

Application of symmetries

Applying a symmetry means transforming a vector or a matrix (already in q-space) into a new vector (matrix). If the vector (matrix) is invariant under that transformation, then that transformation belong to the symmetry group.

To apply the symmetry to a matrix we use the apply_symmetry_vectorq!. For the matrix, we use apply_symmetry_matrixq!. Both these function modify in-place the first argument, storing the result of the transformation there. Note that, since symmetries are stored in crystalline components, both the vector and the matrix must be in crystalline components.

AtomicSymmetries.apply_symmetry_vectorq! — Function

apply_symmetry_vectorq!(target_vector :: AbstractMatrix{Complex{T}}, original_vector :: AbstractMatrix{Complex{T}}, symmetry_operation :: AbstractMatrix{U}, irt :: Vector{Int}, irt_q:: AbstractVector{Int})

Apply the symmetry on the original vector in q space

Parameters

target_vector : The result (modified inplace) (3n x nq)

original_vector : The original vector (3n x nq)

symmetry_operation : The 3x3 symmetry

irt : The atom-atom association by symmetry

irt_q : The q-q association by symmetry

AtomicSymmetries.apply_symmetry_matrixq! — Function

apply_symmetry_matrixq!(target_matrix :: AbstractArray{Complex{T}, 3}, original_matrix :: AbstractArray{Complex{T}, 3}, symmetry_operation :: AbstractMatrix{U}, irt :: Vector{Int}, q_points :: AbstractMatrix{T}; buffer = default_buffer())

Apply the symmetry on the matrix in q space This subroutine assumes the convention that the phase factor is for each supercell, not atoms. In other words, all the atoms coordinates are computed from the same origin of the supercell they are associated with.

Parameters

target_vector : The result (modified inplace) (3n x nq)

original_vector : The original vector (3n x nq)

symmetry_operation : The 3x3 symmetry

irt : The atom-atom association by symmetry

irt_q : The q-q association by symmetry

buffer : The Bumper.jl buffer for caching memory allocations [Optional]

Enforcing symmetries

One of the most useful operation to do is enforce a specific matrix or vector in q-space to satisfy a given symmetry group.

This can be implemented by applying the complete irreducible representation of the symmetry group. Symmetrization of an ent \phi is obtained as

$\Phi = \frac{1}{N}\sum_{i=1}^N S_i(\phi)$

where $S_i$ is the symmetry operation. The two functions performing the symmetrization are symmetrize_matrix_q! and symmetrize_vector_q!. Also in this case, the dynamical matrix must be provided in crystalline coordinates.

To symmetrize vector and matrices already provided in cartesian coordinates, we must use the appropriate subroutines symmetrize_vector_cartesian_q! and symmetrize_matrix_cartesian_q!. These two subroutines correctly convert the vector/matrix in crystal coordinates before applying the symmetries, and then convert the symmetrized result back in cartesian space. They are the most used subroutines to perform symmetrization in q-space, the equivalent of symmetrize_vector! and symmetrize_fc! for real space.

Here the complete API

AtomicSymmetries.symmetrize_vector_q! — Function

symmetrize_vector_q!(target_gamma :: AbstractVector{T}, original_q :: AbstractArray{Complex{T}, 2}, symmetries :: Symmetries, irt_q :: Vector{Vector{Int}}; buffer = default_buffer() where T

Impose the symmetrization of a vector in q space. Since the symmetrization also imposes translational symmetries, the result is always a vector only at gamma.

The symmetrized vector is supposed to be a displacement (so no translations are applied)

NOTE: The provided vector must be in crystal coordinates To symmetrize a vector incartesian coordinates, see the routine symmetrize_vector_cartesian_q!.

Parameters

target_gamma : The n_at * n_dims output symmetrized vector at $\Gamma$

original_q : The original vector in q-space of size nat*n_dims, nq

symmetries : The symmetry group

irt_q : A vector (one for each symmetry) of the correspondances of q points. For each symmetry can be obtained from get_irt_q!

gamma_index : Specify which q vector is $\Gamma$. If not specified, it is assumed to be the first one

AtomicSymmetries.symmetrize_matrix_q! — Function

symmetrize_matrix_q!(target_q :: AbstractArray{Complex{T}, 3}, original_q :: AbstractArray{Complex{T}, 3}, symmetries :: Symmetries, irt_q :: Vector{Vector{Int}}; buffer = default_buffer() where T
symmetrize_matrix_q!(matrix_q :: AbstractArray{Complex{T}, 3}, q_symmetries :: SymmetriesQSpace; buffer = default_buffer())  where T
symmetrize_matrix_q!(target_q :: AbstractArray{T, 3}, original_q :: AbstractArray{Complex{T}, 3}, q_symmetries :: SymmetriesQSpace; buffer = default_buffer())  where T

Impose the symmetrization of a dynamical matrix in q space. The matrix must be in crystal coordinates.

Parameters

target_q : The symmetrized matrix of size n_modes, n_modes, nq (modified in-place).

original_q : The original matrix in q-space of size n_modes, n_modes, nq. It could be the same as target_q

symmetries : The symmetry group

irt_q : A vector (one for each symmetry) of the correspondances of q points. For each symmetry can be obtained from get_irt_q!

minus_q_index : A vector containing for each q the corresponding $\vec {q'} = -\vec q + \vec G$, where $\vec G$ is a generic reciprocal lattice vector.

AtomicSymmetries.symmetrize_vector_cartesian_q! — Function

symmetrize_vector_cartesian_q!(vector_q_cart:: AbstractArray{Complex{T}, 2}, cell :: Matrix{T}, symmetries :: SymmetriesQSpace; buffer = default_buffer()) where {T}

Perform the symmetrization of a vector (overwriting it) in cartesian coordinates. This is the go-to subroutine for performing symmetrization of vectors in q space.

Parameters

vector_q_cart : in-place symmetrize vector (q-space, cartesian coordinates)

cell : 3x3 matrix of the primitive cell (column-based)

symmetries : Symmetries in Q space

buffer : Optional, Bumper stack buffer (caching)

AtomicSymmetries.symmetrize_matrix_cartesian_q! — Function

symmetrize_matrix_cartesian_q!(matrix_q :: AbstractArray{Complex{T}, 3}, cell :: Matrix{T}, q_symmetries :: SymmetriesQSpace; buffer=default_buffer()) where T

Enforce the symmetries on the provided matrix (q-space), modifying it in-place. The provided matrix must be in Cartesian Coordinates.

Parameters

matrix_q : The matrix to be symmetrized. Size (nmodes, nmodes, nq)

cell : The 3x3 primitive cell (column ordering)

q_symmetries : The symmetry group (q-space)

buffer : Optional, stack for Bumper to cache allocations.